Rename classes to singular; edited within code.
