+++ lib/game_stats.rb
@@ -1,35 +1,38 @@
-class GameStats
-  attr_reader :game_id,
-              :team_id,
-              :hoa,
-              :result,
-              :settled_in,
-              :head_coach,
-              :goals,
-              :shots,
-              :tackles,
-              :pim,
-              :powerPlayOpportunities,
-              :powerPlayGoals,
-              :faceOffWinPercentage,
-              :giveaways,
-              :takeaways
+require 'CSV'
 
-  def initialize(stats)
-    @game_id = stats[:game_id]
-    @team_id = stats[:team_id]
-    @hoa = stats[:hoa]
-    @result = stats[:result]
-    @settled_in = stats[:settled_in]
-    @head_coach = stats[:head_coach]
-    @goals = stats[:goals]
-    @shots = stats[:shots]
-    @tackles = stats[:tackles]
-    @pim = stats[:pim]
-    @powerPlayOpportunities = stats[:powerPlayOpportunities]
-    @powerPlayGoals = stats[:powerPlayGoals]
-    @faceOffWinPercentage = stats[:faceOffWinPercentage]
-    @giveaways = stats[:giveaways]
-    @takeaways = stats[:takeaways]
+class GameStats
+    attr_reader :game_id,
+                :team_id,
+                :HoA,
+                :result,
+                :settled_in,
+                :head_coach,
+                :goals,
+                :shots,
+                :tackles,
+                :pim,
+                :powerPlayOpportunities,
+                :powerPlayGoals,
+                :faceOffWinPercentage,
+                :giveaways,
+                :takeaways
+
+    def initialize(data)
+      @game_id = data[:game_id]
+      @team_id = data[:team_id]
+      @HoA = data[:HoA]
+      @result = data[:result]
+      @settled_in = data[:settled_in]
+      @head_coach = data[:head_coach]
+      @goals = data[:goals]
+      @shots = data[:shots]
+      @tackles = data[:tackles]
+      @pim = data[:pim]
+      @powerPlayOpportunities = data[:powerPlayOpportunities]
+      @powerPlayGoals = data[:powerPlayGoals]
+      @faceOffWinPercentage = data[:faceOffWinPercentage]
+      @giveaways = data[:giveaways]
+      @takeaways = data[:takeaways]
+      # binding.pry
+    end
   end
-end
