+++ test/game_test.rb
