+++ test/test_helper.rb
@@ -4,6 +4,6 @@ SimpleCov.start
 require 'minitest/pride'
 require 'minitest/autorun'
 
-require './lib/teams'
-require './lib/games'
+require './lib/team'
+require './lib/game'
 require './lib/game_stats'
